Feat: Confirm deletion for `bm db delete`

+++ src/index.ts
@@ -11,6 +11,8 @@ import {
   getCurrentGitBranch,
   validateOption,
 } from "./util";
+import prompts from "prompts";
+import * as process from "node:process";
 
 const program = new Command();
 
@@ -150,6 +152,7 @@ interface DbDeleteOptions {
   dbName?: string;
   dbUser?: string;
   rootDatabaseUrl?: string;
+  yes?: boolean;
 }
 
 dbCommand
@@ -164,7 +167,20 @@ dbCommand
     "--root-database-url <url>",
     "Root database URL (overrides ROOT_DATABASE_URL env var)"
   )
+  .option('-y, --yes', "Skip confirmation")
   .action(async (options: DbDeleteOptions) => {
+    if (!options.yes) {
+      const { confirm } = await prompts({
+        type: "confirm",
+        name: "confirm",
+        message: `Are you sure you want to delete the database ${options.dbName}?`,
+        initial: false,
+      });
+      if (!confirm) {
+        process.exit(1);
+      }
+    }
+
     await deleteDatabase({
       dbName: getDbName(options),
       dbUser: getDbUserOptional(options),
